add book routes

+++ src/book/book.service.ts
@@ -1,10 +1,11 @@
 /* eslint-disable prettier/prettier */
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Book } from './schema/book.schema';
 import * as mongoose from 'mongoose';
 import { CreateBookDTO } from './dto/CreateBookDTO';
 import { UpdateBookDTO } from './dto/UpdateBookDTO';
+import { Query } from 'express-serve-static-core';
 
 @Injectable()
 export class BookService {
@@ -13,8 +14,22 @@ export class BookService {
         private bookModel: mongoose.Model<Book>
     ) {}
 
-    async findAll(): Promise<Book[]> {
-        const books = await this.bookModel.find();
+    async findAll( query: Query ): Promise<Book[]> {
+
+        const resultPerPage = 2;
+        const currentPage = Number(query.page) || 1;
+        const skip = (currentPage - 1) * resultPerPage
+
+        const keyword = query.keyword ? {
+            title: {
+                $regex: query.keyword,
+                $options: 'i'
+            }
+        } : {};
+     const books = await this.bookModel
+     .find({...keyword})
+     .limit(resultPerPage)
+     .skip(skip);
         return books;
     }
 
@@ -24,14 +39,30 @@ export class BookService {
     }
 
     async findById(id: string): Promise<Book> {
+
+        const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+        if(!isValidId) {
+            throw new BadRequestException('Invalid id used');
+        }
+
         const book = await this.bookModel.findById(id); 
+
         if(!book) {
             throw new NotFoundException('Book not found');
         }
+
         return book
     }
 
     async updateBookById(id: string, book: UpdateBookDTO): Promise<Book> {
+
+        const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+        if(!isValidId) {
+            throw new BadRequestException('Invalid id used');
+        }
+
         const Updatedbook = await this.bookModel.findByIdAndUpdate(id, book, {
             new: true,
             runValidators: true
@@ -44,6 +75,13 @@ export class BookService {
     }
 
     async deleteById(id: string): Promise<string> {
+
+        const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+        if(!isValidId) {
+            throw new BadRequestException('Invalid id used');
+        }
+
         const book = await this.bookModel.findByIdAndDelete(id); 
         if(!book) {
             throw new NotFoundException('Book not found');
